Time: 20 ms (85.12%), Space: 59.2 MB (96.42%) - LeetHub

+++ 2594-minimum-time-to-repair-cars/2594-minimum-time-to-repair-cars.cpp
@@ -4,9 +4,9 @@ public:
         // r*n^2 < time
         // n^2 < time/r
         // n < sqrt(time/r)
-        int rqCars = 0;
+        long long rqCars = 0;
         for(auto &r : ranks){
-            int i = sqrt(time/r);
+            long long i = sqrt(time/r);
             rqCars += i;
             if(rqCars >= cars)   return true;
         }
